working out event page state
