+++ src/components/createpotluckform/CreatePotluckForm.js
@@ -1,7 +1,7 @@
 // Libraries
 import React from 'react';
 import { withFormik, Field, Form } from 'formik';
-import * as Yup from 'yup';
+
 
 const CreatePotluck = () => {
 	return (
