+++ src/components/events/EventGuestList.js
@@ -3,15 +3,16 @@ import React from 'react';
 import { useSelector } from "react-redux";
 
 // Components 
-// import GuestCard from 'components/events/GuestCard';
+import GuestCard from 'components/events/GuestCard';
 
+const EventGuestList = () => {
+	const currentEvent = useSelector(state => state.currentEvent);
 
-const EventGuestList = props => {
 	return (
   <div>
-			{/* {props.map((guest, index) => (
+			{currentEvent.guest? currentEvent.guest.map((guest, index) => (
 				<GuestCard key={index} guest={guest} />
-			))} */}
+			)) : 'Add Guest'}
   </div>
   );
 };
