+++ src/components/events/PotluckInfo.js
@@ -1,16 +1,20 @@
 // Libraries
 import React from "react";
-import { connect } from "react-redux";
 import { useSelector } from "react-redux";
 
-export const PotluckInfo = ({props}) => {
+
+export const PotluckInfo = () => {
+	const currentEvent = useSelector(state => state.currentEvent);
+
 	return (
 		<div>
-      {/* <p>Event Name: {props}</p>
-      <p>Date: {props}</p>
-      <p>Time: {props}</p>
-      <p>Address: {props}</p> */}
+      <p>Event Name: {currentEvent.event_name}</p>
+      <p>Date: {currentEvent.dates}</p>
+      <p>Time: {currentEvent.time}</p>
+      <p>Location: {currentEvent.location}</p>
 
 		</div>
 	);
 };
+
+export default PotluckInfo;
