+++ src/states/events/eventsReducer.js
@@ -1,7 +1,7 @@
 // Types
 import * as types from "states/events/eventsTypes";
-
-export const eventsReducers = (events = {}, action) => {
+const intialEventState = {events:[], currentEvent:{}}
+export const eventsReducers = (events = intialEventState, action) => {
   switch (action.type) {
     case types.GET_EVENT_INFO:
       return action.payload;
